+++ tests/wildboar/conftest.py
@@ -1,11 +1,27 @@
 import numpy as np
 import pytest
-from wildboar.datasets import load_gun_point
+
+from wildboar.datasets import load_ering, load_gun_point
 
 
 @pytest.fixture(scope="session")
 def gun_point():
     return load_gun_point(merge_train_test=False)
+
+
+@pytest.fixture(scope="session")
+def ering():
+    return load_ering(merge_train_test=False)
+
+
+@pytest.fixture(scope="session")
+def load_data(request):
+    if request.param == "uts":
+        return load_gun_point()
+    elif request.param == "mts":
+        return load_ering()
+    else:
+        raise ValueError("unsupported data")
 
 
 @pytest.fixture(scope="session")